+++ lib/pages/my_page/not_login_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:project_finaltask/pages/toppage.dart';
+import 'package:project_finaltask/pages/top_page.dart';
 import 'package:project_finaltask/utils/color_extension.dart';
 
 class NotLoginView extends StatelessWidget {
