+++ lib/pages/settings_page/privacy_policy_screen.dart
