+++ lib/pages/settings_page/tos_screen.dart
