+++ lib/pages/user_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:project_finaltask/components/appbar.dart';
+import 'package:project_finaltask/components/appbar_component.dart';
 
 import '../../models/user.dart';
-import '../views/userview.dart';
+import '../views/use_rview.dart';
 
 class UserPage extends StatefulWidget {
   late final User user;
