+++ lib/views/user_list_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 
 import '../models/user.dart';
-import '../pages/userpage.dart';
+import '../pages/user_page.dart';
 import '../utils/color_extension.dart';
 
 class UserListView extends StatelessWidget {
